# src/main/java/com/example/finalproject/model/dto/response/ClassRoomResponse.java
package com.example.finalproject.model.dto.response;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ClassRoomResponse {
    private Long id;
    private String classCode;
    private String className;
    private String description;
    private String status;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}